feat: move nav, add console mock
