+++ frontend/src/app/app.scss
@@ -22,15 +22,6 @@ header {
   }
 }
 
-nav {
-  display: flex;
-  flex-direction: row;
-  padding: 10px 16px;
-  justify-content: space-between;
-  gap: 12px;
-  width: 100%;
-}
-
 main {
   flex: 1;
   display: flex;
